# docs/api.yaml
openapi: 3.0.0
info:
  title: Pet Socialization Service Platform API
  description: 宠物社交服务平台接口文档
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: 本地开发服务器
paths:
...
  /api/posts:
    get:
      summary: 获取帖子列表
      tags: [Posts]
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Post'

    post:
      summary: 创建帖子
      tags: [Posts]
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PostCreate'
      responses:
        '201':
          description: 创建成功
  /api/posts/{id}:
    delete:
      summary: 删除帖子
      tags: [Posts]
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: 删除成功
...
  /api/services/{id}/reviews:
    get:
      summary: 获取商户评论
      tags: [ServiceReviews]
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: 成功


    post:
      summary: 创建商户评论
      tags: [ServiceReviews]
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceReviewCreate'
      responses:
        '201':
          description: 评论成功
...
components:
  schemas:
...
    Post:
      type: object
      properties:
        post_id:
          type: integer
        user_id:
          type: integer
        content:
          type: string
        images:
          type: array
          items:
            type: string
        like_count:
          type: integer
        comment_count:
          type: integer
        create_time:
          type: string
          format: date-time
    PostCreate:
      type: object
      properties:
        content:
          type: string
        images:
          type: array
          items:
            type: string
...
    ServiceReviewCreate:
      type: object
      properties:
        rating:
          type: number
        content:
          type: string
...